Code not working. Revert

+++ core/engine/schemes/convection.c
@@ -27,7 +27,7 @@ void CFD_Setup_Convection(CFD_t *cfd)
     }
 }
 
-void CFD_Scheme_Get_Flux(CFD_t *cfd, uint16_t i, uint16_t j, phi_t phi)
+void CFD_Scheme_Get_Flux(CFD_t *cfd, int i, int j, phi_t phi)
 {
 
     double uPP = CFD_Get_State(cfd, u, i + 0, j + 0);
